+++ src/main/java/com/sealog/backend/domain/feature/user/controller/GuestUserController.java
@@ -8,22 +8,23 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api/user")
+@RequestMapping("/api/guest")
 @RequiredArgsConstructor
-public class BlogInfoController implements BlogInfoControllerDocs{
+public class GuestUserController implements GuestUserControllerDocs {
 
     private final UserService userService;
 
     /**
      * 블로그 사용자 정보
-     * GET /api/user/{nickname}
+     * GET /api/guest/{nickname}
+     * 게스트
      */
     @Override
     @GetMapping("/{nickname}")
-    public ResponseEntity<CustomResponse<UserResponse.BlogUserInfo>> signUp(
+    public ResponseEntity<CustomResponse<UserResponse.BlogUserInfo>> getBlogUserInfo(
             @PathVariable String nickname
     ) {
         UserResponse.BlogUserInfo blogUser = userService.getBlogUser(nickname);
-        return ResponseEntity.ok(CustomResponse.success(blogUser, "회원가입이 성공적으로 완료되었습니다."));
+        return ResponseEntity.ok(CustomResponse.success(blogUser, "정보 조회가 성공적으로 완료되었습니다."));
     }
 }
