+++ src/main/java/com/sealog/backend/domain/feature/user/controller/GuestUserControllerDocs.java
@@ -13,7 +13,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 
 @Tag(name = "Blog", description = "블로그 사용자 정보 API (인증 불필요)")
-public interface BlogInfoControllerDocs {
+public interface GuestUserControllerDocs {
 
     @Operation(summary = "블로그 사용자 정보 조회", description = "닉네임으로 블로그 사용자 정보를 조회합니다.")
     @SecurityRequirements()
@@ -22,7 +22,7 @@ public interface BlogInfoControllerDocs {
             @ApiResponse(responseCode = "404", description = "사용자 없음",
                     content = @Content(schema = @Schema(hidden = true))),
     })
-    ResponseEntity<CustomResponse<UserResponse.BlogUserInfo>> signUp(
+    ResponseEntity<CustomResponse<UserResponse.BlogUserInfo>> getBlogUserInfo(
             @Parameter(description = "사용자 닉네임", example = "테스터") String nickname
     );
 }
